refactor: 에러 메세지 상수화 및 공용 confirm 모달 생성 후 적용
